working on dungeon generation, simplified bsptree file
