+++ src/main/scala/game_engine/pcg/BSPTree.scala
@@ -1,21 +1,15 @@
 package game_engine.pcg
 
-import game_object_system.graphics_objects.Rect
-
-import scala.util.Random
+import DungeonGenerator.random
+import DungeonGenerator.minRoom
 
 sealed trait BSPTree
 case class BSPNode(x : Int, y : Int, w: Int, h : Int, left : BSPTree, right: BSPTree) extends BSPTree
 case class BSPLeaf(x : Int, y : Int, w: Int, h : Int) extends BSPTree
 
-object BSPBuilder {
+object BSPTree {
 
-  val random: Random = Random
-  val tileSize = 16
-  val minRoom: Int = 3 * tileSize
-
-
-  def buildDungeon(limit : Int, width : Int, height : Int) : BSPTree = {
+  def buildBSPTree(limit : Int, width : Int, height : Int) : BSPTree = {
     def buildBSP(l: Int, x: Int, y: Int, w: Int, h: Int): BSPTree = {
       if (l >= limit) BSPLeaf(x, y, w, h)
       else {
@@ -46,12 +40,6 @@ object BSPBuilder {
 
     buildBSP(0, 0, 0, width, height)
   }
-  def createRooms(t : BSPTree) : List[Rect] = t match {
-    case BSPNode(_, _, _, _, l, r) => createRooms(l) ++ createRooms(r)
-    case BSPLeaf(x,y,w,h) =>
-      val rw = Math.abs(random.nextInt()) % (w-minRoom+1)
-      val rh = Math.abs(random.nextInt()) % (h-minRoom+1)
-      List(Rect(x+rw,y+rh,w,h))
-  }
+
   private def split(r : Int) = random.between(minRoom, r-minRoom)
 }
